requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
pandas==1.5.3
requests==2.28.1
bcrypt==4.0.1
Pillow==9.5.0
gunicorn
psycopg2-binary
python-dotenv